Added: Started on functionality on /user/:user_id

+++ server/server/src/main/java/se/experis/tidsbanken/server/repositories/CommentRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import javax.xml.stream.events.Comment;
 
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    Comment getById(int id);
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    Comment getById(long id);
 }
